+++ Anarchy/REST/Guild/Role/Models/DiscordRole.cs
@@ -31,14 +31,7 @@ namespace Discord
 
 
         [JsonProperty("permissions")]
-#pragma warning disable IDE0051, IDE1006
-        private uint _permissions
-        {
-            get { return Permissions;  }
-            set { Permissions = new DiscordPermissions(value); }
-        }
-#pragma warning restore IDE0051, IDE1006
-        public DiscordPermissions Permissions { get; private set; }
+        public DiscordPermission Permissions { get; private set; }
 
 
         /// <summary>