+++ Anarchy/REST/Permissions/OverwrittenPermissionState.cs
@@ -1,0 +1,9 @@
+﻿namespace Discord
+{
+    public enum OverwrittenPermissionState
+    {
+        Allow,
+        Inherit,
+        Deny
+    }
+}