+++ Anarchy/Voice/DiscordVoiceSession.cs
@@ -176,8 +176,6 @@ namespace Discord.Voice
                         State = DiscordVoiceClientState.Connected;
 
                         OnConnected?.Invoke(this, null);
-
-                        StartSpeakAsync();
                         break;
                     case DiscordVoiceOpcode.Hello:
                         var ident = new DiscordVoiceIdentify()
@@ -201,17 +199,6 @@ namespace Discord.Voice
                         catch { }
 
                         break;
-                }
-            });
-        }
-
-        private async void StartSpeakAsync()
-        {
-            await Task.Run(() =>
-            {
-                while (true)
-                {
-                    // TODO: figure out a way to detect when the user has stopped speaking
                 }
             });
         }